even better

+++ minifi-native/src/c_ffi/c_ffi_controller_service_context.rs
@@ -33,10 +33,8 @@ unsafe extern "C" fn property_callback(
             return;
         }
 
-        let value_slice = std::slice::from_raw_parts(
-            property_c_value.data as *const u8,
-            property_c_value.length as usize,
-        );
+        let value_slice =
+            std::slice::from_raw_parts(property_c_value.data as *const u8, property_c_value.length);
         if let Ok(string_value) = String::from_utf8(value_slice.to_vec()) {
             *result_target = Some(string_value);
         }
